[Util] Encapsulate file name attributes functions into CFilenameAttribues
