+++ xbmc/Util.cpp
@@ -89,6 +89,7 @@
 #include "utils/LangCodeExpander.h"
 #include "utils/RegExp.h"
 #include "utils/StringUtils.h"
+#include "video/FilenameAttributes.h"
 #include "video/VideoDatabase.h"
 #include "video/VideoFileItemClassify.h"
 #include "windowing/GraphicContext.h"
@@ -103,16 +104,13 @@
 #include <chrono>
 #include <cstdint>
 #include <cstdlib>
-#include <functional>
 #include <iomanip>
 #include <memory>
 #include <random>
 #include <ranges>
 #include <sstream>
 #include <string>
-#include <string_view>
 #include <unordered_set>
-#include <utility>
 #include <vector>
 
 #include <fstrcmp.h>
@@ -128,8 +126,6 @@ using namespace KODI;
 #if !defined(TARGET_WINDOWS)
 unsigned int CUtil::s_randomSeed = time(NULL);
 #endif
-
-constexpr std::string_view FILENAME_EDITION = "edition";
 
 namespace
 {
@@ -453,182 +449,6 @@ std::string CUtil::GetPartNumberFromPath(std::string path)
   return GetPartAndRemoveDiscFromPath(path, PreserveFileName::REMOVE);
 }
 
-namespace
-{
-/*!
- * \brief Returns a compiled regular expression to retrieve filename attributes
- * \param[in] cache Optional regular expression cache
- * \return Valid pointer if the regular expression was compiled successfully, nullptr otherwise.
- */
-std::shared_ptr<CRegExp> InitFilenameAttributesRegExp(KODI::REGEXP::RegExpCache* cache)
-{
-  std::shared_ptr<CRegExp> re;
-
-  const std::shared_ptr<CAdvancedSettings> advancedSettings =
-      CServiceBroker::GetSettingsComponent()->GetAdvancedSettings();
-
-  if (advancedSettings == nullptr)
-    return re;
-
-  re = KODI::REGEXP::GetRegExp(advancedSettings->m_videoFilenameAttributePairsRegExp, cache, true,
-                               CRegExp::autoUtf8);
-  if (re == nullptr)
-  {
-    CLog::LogF(LOGERROR, "Invalid filename attribute pairs RegExp:'{}'",
-               advancedSettings->m_videoFilenameAttributePairsRegExp);
-  }
-  return re;
-}
-
-/*!
- * \brief Iterates over all filename attributes key=value pairs in @p fileName, invoking @p callback
- *        for each one.
- * @p fileName must not be modified by @p callback.
- * \param fileName [in] The filename to search for attribute pairs.
- * \param cache[in] Optional regular expression cache
- * \param callback[in] Invoked for each match with the absolute position, the match length, the value
- *                     of the key and value of the match.
- */
-void ForEachFilenameAttribute(
-    const std::string& fileName,
-    KODI::REGEXP::RegExpCache* cache,
-    std::function<void(int pos, int len, const std::string& key, const std::string& value)>
-        callback)
-{
-  if (std::shared_ptr<CRegExp> re = InitFilenameAttributesRegExp(cache); re != nullptr)
-  {
-    unsigned int offset = 0;
-    while (true)
-    {
-      int pos = re->RegFind(fileName, offset);
-      if (pos < 0)
-        break;
-
-      const int matchLength = re->GetFindLen();
-      callback(pos, matchLength, re->GetMatch("key"), re->GetMatch("value"));
-      offset = pos + matchLength;
-    }
-  }
-}
-} // namespace
-
-CUtil::FilenameAttributeMap CUtil::GetFilenameAttributePairs(const std::string& fileName,
-                                                             KODI::REGEXP::RegExpCache* cache)
-{
-  FilenameAttributeMap result;
-
-  ForEachFilenameAttribute(fileName, cache,
-                           [&result](int, int, std::string key, std::string value)
-                           {
-                             StringUtils::Trim(key);
-                             StringUtils::ToLower(key);
-                             StringUtils::Trim(value);
-                             if (!key.empty() && !value.empty())
-                               result.insert_or_assign(key, value);
-                           });
-  return result;
-}
-
-void CUtil::CleanFilenameAttributePairs(std::string& fileName, KODI::REGEXP::RegExpCache* cache)
-{
-  std::string result;
-  result.reserve(fileName.size());
-  int last = 0;
-
-  // Collect the gaps between attribute tokens into the new result string. Reduces allocations
-  // and in place erasure without modification of the string being iterated.
-  ForEachFilenameAttribute(
-      fileName, cache,
-      [&result, &fileName, &last](int pos, int len, const std::string&, const std::string&)
-      {
-        result.append(fileName, last, pos - last);
-        last = pos + len;
-      });
-  result.append(fileName, last, std::string::npos);
-
-  fileName = std::move(result);
-}
-
-bool CUtil::GetFilenameIdentifier(const std::string& fileName,
-                                  std::string& identifierType,
-                                  std::string& identifier,
-                                  KODI::REGEXP::RegExpCache* cache)
-{
-  return GetFilenameIdentifier(GetFilenameAttributePairs(fileName, cache), identifierType,
-                               identifier);
-}
-
-bool CUtil::GetFilenameIdentifier(const CUtil::FilenameAttributeMap& attributes,
-                                  std::string& identifierType,
-                                  std::string& identifier)
-{
-  if (const std::shared_ptr<CAdvancedSettings> advancedSettings =
-          CServiceBroker::GetSettingsComponent()->GetAdvancedSettings();
-      advancedSettings != nullptr)
-  {
-    const std::unordered_set<std::string>& identifiers =
-        advancedSettings->m_videoScannerMetadataSources;
-
-    if (identifiers.empty())
-      return false;
-
-    // The attribute key must be a known identifier with an optional id suffix and the value
-    // must be purely alphanumeric.
-
-    // Projection to strip the "id" suffix
-    auto proj = [](const std::pair<std::string, std::string>& attr)
-    {
-      constexpr std::string_view suffix = "id";
-      if (!StringUtils::EndsWithNoCase(attr.first, suffix))
-        return attr;
-      return std::pair{attr.first.substr(0, attr.first.size() - suffix.size()), attr.second};
-    };
-
-    auto it = std::ranges::find_if(
-        attributes,
-        [&identifiers](const auto& attr)
-        {
-          if (identifiers.contains(attr.first) && !attr.second.empty() &&
-              std::ranges::all_of(attr.second,
-                                  [](char c) { return StringUtils::isasciialphanum(c); }))
-            return true;
-
-          return false;
-        },
-        proj);
-
-    if (it != attributes.end())
-    {
-      auto stripped = proj(*it);
-      identifierType = stripped.first;
-      identifier = stripped.second;
-      return true;
-    }
-  }
-  return false;
-}
-
-bool CUtil::HasFilenameIdentifier(const std::string& fileName)
-{
-  std::string identifierType;
-  std::string identifier;
-  return GetFilenameIdentifier(fileName, identifierType, identifier, nullptr);
-}
-
-std::string CUtil::GetFilenameEdition(const std::string& fileName, KODI::REGEXP::RegExpCache* cache)
-{
-  return GetFilenameEdition(GetFilenameAttributePairs(fileName, cache));
-}
-
-std::string CUtil::GetFilenameEdition(const FilenameAttributeMap& attributes)
-{
-  auto it = attributes.find(FILENAME_EDITION);
-  if (it != attributes.end())
-    return it->second;
-
-  return "";
-}
-
 void CUtil::CleanString(const std::string& strFileName,
                         std::string& strTitle,
                         std::string& strTitleAndYear,
@@ -641,7 +461,7 @@ void CUtil::CleanString(const std::string& strFileName,
   if (strFileName == "..")
    return;
 
-  CleanFilenameAttributePairs(strTitleAndYear, nullptr);
+  KODI::VIDEO::CFilenameAttributes::CleanFilenameAttributePairs(strTitleAndYear, nullptr);
 
   const std::shared_ptr<CAdvancedSettings> advancedSettings = CServiceBroker::GetSettingsComponent()->GetAdvancedSettings();
   const std::vector<std::string> &regexps = advancedSettings->m_videoCleanStringRegExps;
