+++ xbmc/video/jobs/VideoLibraryRefreshingJob.cpp
@@ -32,6 +32,7 @@
 #include "utils/StringUtils.h"
 #include "utils/URIUtils.h"
 #include "utils/log.h"
+#include "video/FilenameAttributes.h"
 #include "video/VideoDatabase.h"
 #include "video/VideoInfoDownloader.h"
 #include "video/VideoInfoScanner.h"
@@ -292,7 +293,7 @@ bool CVideoLibraryRefreshingJob::Work(CVideoDatabase &db)
       CVideoInfoDownloader infoDownloader(scraper);
 
       // try adding by filename identifier
-      if (scraper->IsPython() && CUtil::HasFilenameIdentifier(itemTitle))
+      if (scraper->IsPython() && CFilenameAttributes(itemTitle, nullptr).HasIdentifier())
       {
         CFileItemList items;
         items.Add(m_item);
